My name is Matthew Rosan I have a background in SAAS sales and Nasdaq futures trading.
My career goal is to become a quantitative trader.

I chose Fintech because I know as a retail trader that manual traders are a dying breed and if I want to stay ahead of the curve I need to learn a new set of skills to better position myself in the marketplace.

Updated*